requirements.txt
# --- 既存のライブラリ ---
streamlit
pandas
plotly
python-dateutil
openpyxl
streamlit-oauth
gspread
PyYAML
llama-cpp-python
requests

# --- 今回の改修で追加するライブラリ ---
sentence-transformers
faiss-cpu